openaiservice/requirements.txt
fastapi
requests
uvicorn
chainlit
chromadb
langchain
langchain_openai
openai
pypdf
python-dotenv
tiktoken

